Board Layout
Implemented a button to call a canvas of the board, currently lays out a honey comb structure but it is hard to tell whether there is overlap or not.
Next step is changing the tile to a empty hex with a white outline to create the back of the board. And then  have positions for when tiles are placed by the user.